pyproject.toml
[tool.poetry]
name = "ai-doc-generator"
version = "0.1.0"
description = "AI-powered documentation generator for any software repository"
authors = ["AI Doc Generator <docs@example.com>"]
readme = "README_PROJECT.md"
packages = [
    { include = "api" },
    { include = "core" },
    { include = "parser" },
    { include = "analyzer" },
    { include = "chunking" },
    { include = "vector_store" },
    { include = "generator" },
    { include = "cli" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
tree-sitter = "^0.21.3"
tree-sitter-languages = "^1.10.2"
gitpython = "^3.1.43"
networkx = "^3.3"
faiss-cpu = "^1.8.0"
tqdm = "^4.66.4"
pydantic = "^2.7.1"
pydantic-settings = "^2.3.0"
click = "^8.1.7"
rich = "^13.7.1"
python-dotenv = "^1.0.1"
numpy = "^1.26.4"
httpx = "^0.28.1"
google-genai = "^1.67.0"
types-protobuf = "^6.32.1.20260221"
tenacity = "^8.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
docgen = "cli.generate_docs:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
